♻️ Refactor: Rename Auth feature to User
